funções Handle na classe Handler
